incorporate Lexeme Details into Lexeme Page
