+++ pages/Lexeme/Lexeme.less
@@ -1,19 +1,106 @@
 @import 'Label/Label.less';
-@import 'Lexeme/LexemeDetails.less';
 @import 'Numbered/Numbered.less';
 @import 'PageTitle/PageTitle.less';
 @import 'Translation/Translation.less';
 @import 'MultiOrthoTranscription/MultiOrthoTranscription.less';
 
+// Handle case where the URL has no hash.
+body:target #form {
+  display: flex;
+}
+
 .lexeme-page {
+
+  @breakpoint: 45em;
 
   align-self     : center;
   inline-size    : 100%;
   max-inline-size: var(--measure);
 
-  .error-message {
-    margin    : var(--space);
-    text-align: center;
+  .label {
+    padding: var(--text-padding);
+  }
+
+  .mot {
+    padding: var(--text-padding);
+  }
+
+  nav {
+    margin-block-start: var(--space);
+  }
+
+  .page-title {
+    align-self : flex-start;
+    font-family: var(--font-unicode);
+    text-align : left;
+  }
+
+  .panel:not(.current, :target) {
+    display: none;
+  }
+
+  .tab:hover {
+    background-color: var(--cyan-1);
+  }
+
+  .tab-link {
+    padding: var(--text-padding-lg);
+  }
+
+  .tabs {
+    border-bottom: var(--border);
+    font-size    : large;
+    inline-size  : 100%;
+  }
+
+  @media (prefers-color-scheme: dark) {
+    .tab:hover {
+      color: var(--background-color-dark);
+    }
+  }
+
+  // small screens
+  @media (max-width: @breakpoint) {
+
+    [aria-selected] {
+      position: relative;
+    }
+
+    [aria-selected]::before {
+      background-color  : var(--cyan);
+      inset-block-end   : 0;
+      content           : '';
+      inset-inline-start: 0;
+      position          : absolute;
+      inset-block-start : 0;
+      inline-size       : 0.25em;
+    }
+
+    .tab:first-child {
+      border-block-start: var(--border);
+    }
+
+    .tab:not(:last-child) {
+      border-block-end: var(--border);
+    }
+
+  }
+
+  // large screens
+  @media (min-width: @breakpoint) {
+
+    [aria-selected] {
+      border-bottom: 0.25em solid var(--cyan);
+    }
+
+    .page-title {
+      padding-inline: 0;
+    }
+
+    .tabs {
+      flex-direction: row;
+    }
+
   }
 
 }
