--- pages/Lexeme/LexemeDetails.hbs
@@ -1,33 +1,0 @@
-<section class=lexeme-details>
-
-  <header>
-    <h1 class=page-title lang='{{ language.iso }}'>{{mot lexeme.lemma language.defaultOrthography }}</h1>
-    <ol class=numbered>
-      {{#each lexeme.senses as |sense| }}
-        <li class=tln>{{mls sense.gloss }}</li>
-      {{/each}}
-    </ol>
-  </header>
-
-  <nav>
-    <ul class=tabs role=tablist>
-      <li role=tab class=tab><a class=tab-link id=form-link href=#form aria-selected=true>Form</a></li>
-      <li role=tab class=tab><a class=tab-link id=meaning-link href=#meaning>Meaning</a></li>
-      <li role=tab class=tab><a class=tab-link id=metadata-link href=#metadata>Metadata</a></li>
-    </ul>
-  </nav>
-
-  <section class='current panel' role=tabpanel id=form aria-labelledby=form-link>
-    <label class=label>Lemma</label>
-    {{> MultiOrthoTranscription/MultiOrthoTranscription txn=lexeme.lemma lang=language.iso }}
-  </section>
-
-  <section class=panel role=tabpanel id=meaning aria-labelledby=meaning-link>
-    <h1>Meaning</h1>
-  </section>
-
-  <section class=panel role=tabpanel id=metadata aria-labelledby=metadata-link>
-    <h1>Metadata</h1>
-  </section>
-
-</section>
