--- pages/Lexeme/LexemeDetails.less
@@ -1,101 +1,0 @@
-// Handle case where the URL has no hash.
-body:target #form {
-  display: flex;
-}
-
-.lexeme-details {
-
-  @breakpoint: 45em;
-
-  --selected-width: 0.25em;
-
-  inline-size    : 100%;
-  max-inline-size: var(--measure);
-
-  .label {
-    padding: var(--text-padding);
-  }
-
-  nav {
-    margin-block-start: var(--space);
-  }
-
-  .page-title {
-    align-self : flex-start;
-    font-family: var(--font-unicode);
-    text-align : left;
-  }
-
-  .panel:not(.current, :target) {
-    display: none;
-  }
-
-  .tab:hover {
-    background-color: var(--cyan-1);
-  }
-
-  .tab-link {
-    padding: var(--text-padding-lg);
-  }
-
-  .tabs {
-    border-bottom: var(--border);
-    font-size    : large;
-    inline-size  : 100%;
-  }
-
-  .mot {
-    padding: var(--text-padding);
-  }
-
-  @media (prefers-color-scheme: dark) {
-    .tab:hover {
-      color: var(--background-color-dark);
-    }
-  }
-
-  // small screens
-  @media (max-width: @breakpoint) {
-
-    [aria-selected] {
-      position: relative;
-    }
-
-    [aria-selected]::before {
-      background-color  : var(--cyan);
-      inset-block-end   : 0;
-      content           : '';
-      inset-inline-start: 0;
-      position          : absolute;
-      inset-block-start : 0;
-      inline-size       : 0.25em;
-    }
-
-    .tab:first-child {
-      border-block-start: var(--border);
-    }
-
-    .tab:not(:last-child) {
-      border-block-end: var(--border);
-    }
-
-  }
-
-  // large screens
-  @media (min-width: @breakpoint) {
-
-    [aria-selected] {
-      border-bottom: var(--selected-width) solid var(--cyan);
-    }
-
-    .page-title {
-      padding-inline: 0;
-    }
-
-    .tabs {
-      flex-direction: row;
-    }
-
-  }
-
-}
